Added support for adding comments against a "resource", it uses the controller name in the url, and the passed param (this in most cases will be the id of the record your looking at) I've tried to make this robust as possible, but open to improvment as the method I mentioned above using the controller_name+row id may in a edge case result in comments being logged aginst some random combination.
There is a comments index located at /active_admin/admin_comments so you can see all comments
a link is prodvided to get you back to the item/resource owning the comments.
Author defaults to Admin, because I couldn't work out a reliable way of retrieving the username(differs based on Auth system implmented by the user)

There are predefined filters on the comments index, so you can search comments, by content or by class it came from.

There are options to disable commenting, app wide via ActiveAdmin bootstrap setting
or on a per model basis. Simply add $aaNoComment = true; to models which commenting should be disabled

Finally, data integrity, add this to your app's AppModel, this should remove comments assoicated with a item/resource when it gets deleted
  public function afterDelete()
    {
        //If the active_admin plugin is loaded remove any comments
        if (CakePlugin::loaded("ActiveAdmin")) {
            App::uses('AdminComment', 'ActiveAdmin.Model');
            $aa_comment = new AdminComment();
            $aa_comment->removeComments($this->alias, $this->id);
        }
        return true;
    }

Again this implementation is not perfect, and in my app I have little use for it at current, but it's nice to have some feature like it